+++ src/ConnectionManager.cpp
@@ -10,6 +10,8 @@
 #include <algorithm>
 
 #include <core_pins.h>
+#include <lwip/ip.h>
+
 #include "ConnectionHolder.h"
 #include "QNEthernet.h"
 
@@ -261,7 +263,7 @@ std::shared_ptr<ConnectionHolder> ConnectionManager::connect(ip_addr_t *ipaddr,
   return holder;
 }
 
-bool ConnectionManager::listen(uint16_t port) {
+bool ConnectionManager::listen(uint16_t port, bool reuse) {
   // Try to allocate
   tcp_pcb *pcb = tcp_new();
   if (pcb == nullptr) {
@@ -269,6 +271,9 @@ bool ConnectionManager::listen(uint16_t port) {
   }
 
   // Try to bind
+  if (reuse) {
+    ip_set_option(pcb, SOF_REUSEADDR);
+  }
   if (tcp_bind(pcb, IP_ADDR_ANY, port) != ERR_OK) {
     tcp_abort(pcb);
     return false;
@@ -290,7 +295,7 @@ bool ConnectionManager::listen(uint16_t port) {
   return true;
 }
 
-bool ConnectionManager::isListening(uint16_t port) {
+bool ConnectionManager::isListening(uint16_t port) const {
   auto it = std::find_if(
       listeners_.begin(), listeners_.end(), [port](const auto &elem) {
         return (elem != nullptr) && (elem->local_port == port);
@@ -312,7 +317,7 @@ bool ConnectionManager::stopListening(uint16_t port) {
 }
 
 std::shared_ptr<ConnectionHolder> ConnectionManager::findConnected(
-    uint16_t port) {
+    uint16_t port) const {
   auto it = std::find_if(
       connections_.begin(), connections_.end(), [port](const auto &elem) {
         const auto &state = elem->state;
@@ -326,7 +331,7 @@ std::shared_ptr<ConnectionHolder> ConnectionManager::findConnected(
 
 // TODO: Should we implement some kind of fairness?
 std::shared_ptr<ConnectionHolder> ConnectionManager::findAvailable(
-    uint16_t port) {
+    uint16_t port) const {
   auto it = std::find_if(
       connections_.begin(), connections_.end(), [port](const auto &elem) {
         const auto &state = elem->state;
@@ -340,7 +345,8 @@ std::shared_ptr<ConnectionHolder> ConnectionManager::findAvailable(
   return nullptr;
 }
 
-bool ConnectionManager::remove(const std::shared_ptr<ConnectionHolder> &holder) {
+bool ConnectionManager::remove(
+    const std::shared_ptr<ConnectionHolder> &holder) {
   auto it =
       std::find_if(connections_.begin(), connections_.end(),
                    [holder](const auto &elem) { return (elem == holder); });