+++ src/QNEthernetClient.cpp
@@ -8,13 +8,16 @@
 
 // C++ includes
 #include <algorithm>
+#include <cstring>
 
 #include <elapsedMillis.h>
 #include <lwip/dns.h>
 #include <lwip/netif.h>
 #include <lwip/opt.h>
 #include <lwip/tcp.h>
+
 #include "ConnectionManager.h"
+#include "QNDNSClient.h"
 #include "QNEthernet.h"
 
 namespace qindesign {
@@ -24,28 +27,10 @@ namespace network {
 static constexpr uint32_t kDNSLookupTimeout =
     DNS_MAX_RETRIES * DNS_TMR_INTERVAL;
 
-void EthernetClient::dnsFoundFunc(const char *name, const ip_addr_t *ipaddr,
-                                  void *callback_arg) {
-  if (callback_arg == nullptr || ipaddr == nullptr) {
-    return;
-  }
-
-  EthernetClient *client = reinterpret_cast<EthernetClient *>(callback_arg);
-
-  // Also check the host name in case there was some previous request pending
-  if (client->lookupHost_ == name) {
-    client->lookupIP_ = ipaddr->addr;
-    client->lookupFound_ = true;
-  }
-}
-
 EthernetClient::EthernetClient() : EthernetClient(nullptr) {}
 
 EthernetClient::EthernetClient(std::shared_ptr<ConnectionHolder> conn)
     : connTimeout_(1000),
-      lookupHost_{},
-      lookupIP_{INADDR_NONE},
-      lookupFound_(false),
       conn_(conn) {}
 
 EthernetClient::~EthernetClient() {
@@ -82,28 +67,11 @@ int EthernetClient::connect(IPAddress ip, uint16_t port) {
 }
 
 int EthernetClient::connect(const char *host, uint16_t port) {
-  ip_addr_t addr;
-  lookupHost_ = host;
-  lookupIP_ = INADDR_NONE;
-  lookupFound_ = false;
-  switch (dns_gethostbyname(host, &addr, &dnsFoundFunc, this)) {
-    case ERR_OK:
-      return connect(addr.addr, port);
-    case ERR_INPROGRESS: {
-      elapsedMillis timer;
-      while (lookupIP_ == INADDR_NONE && timer < kDNSLookupTimeout) {
-        // NOTE: Depends on Ethernet loop being called from yield()
-        yield();
-      }
-      if (lookupFound_) {
-        return connect(lookupIP_, port);
-      }
-      return false;
-    }
-    case ERR_ARG:
-    default:
-      return false;
+  IPAddress ip;
+  if (!DNSClient::getHostByName(host, ip, kDNSLookupTimeout)) {
+    return false;
   }
+  return connect(ip, port);
 }
 
 uint8_t EthernetClient::connected() {
@@ -114,6 +82,7 @@ uint8_t EthernetClient::connected() {
     conn_ = nullptr;
     return false;
   }
+  EthernetClass::loop();  // Allow information to come in
   return true;
 }
 
@@ -127,6 +96,7 @@ EthernetClient::operator bool() {
     }
     return false;
   }
+  EthernetClass::loop();  // Allow information to come in
   return true;
 }
 
@@ -150,7 +120,7 @@ void EthernetClient::stop() {
   if (conn_->connected) {
     // First try to flush any data
     tcp_output(state->pcb);
-    EthernetClass::loop();  // Maybe some data gets out
+    EthernetClass::loop();  // Maybe some TCP data gets in
     // NOTE: loop() requires a re-check of the state
 
     if (state != nullptr) {
@@ -169,8 +139,54 @@ void EthernetClient::stop() {
   conn_ = nullptr;
 }
 
+void EthernetClient::close() {
+  if (conn_ == nullptr) {
+    return;
+  }
+
+  const auto &state = conn_->state;
+  if (state == nullptr) {
+    conn_ = nullptr;
+    return;
+  }
+
+  if (conn_->connected) {
+    // First try to flush any data
+    tcp_output(state->pcb);
+    EthernetClass::loop();  // Maybe some TCP data gets in
+    // NOTE: loop() requires a re-check of the state
+
+    if (state != nullptr) {
+      if (tcp_close(state->pcb) != ERR_OK) {
+        tcp_abort(state->pcb);
+      }
+    }
+  }
+
+  conn_ = nullptr;
+}
+
+void EthernetClient::closeOutput() {
+  if (!static_cast<bool>(*this)) {
+    return;
+  }
+  const auto &state = conn_->state;
+  if (state == nullptr) {
+    return;
+  }
+
+  // First try to flush any data
+  tcp_output(state->pcb);
+  EthernetClass::loop();  // Maybe some TCP data gets in
+  // NOTE: loop() requires a re-check of the state
+
+  if (state != nullptr) {
+    tcp_shutdown(state->pcb, 0, 1);
+  }
+}
+
 uint16_t EthernetClient::localPort() {
-  if (!(*this)) {
+  if (!static_cast<bool>(*this)) {
     return 0;
   }
 
@@ -182,7 +198,7 @@ uint16_t EthernetClient::localPort() {
 }
 
 IPAddress EthernetClient::remoteIP() {
-  if (!(*this)) {
+  if (!static_cast<bool>(*this)) {
     return INADDR_NONE;
   }
 
@@ -194,7 +210,7 @@ IPAddress EthernetClient::remoteIP() {
 }
 
 uint16_t EthernetClient::remotePort() {
-  if (!(*this)) {
+  if (!static_cast<bool>(*this)) {
     return 0;
   }
 
@@ -209,8 +225,31 @@ uint16_t EthernetClient::remotePort() {
 //  Transmission
 // --------------------------------------------------------------------------
 
+void EthernetClient::writeFully(uint8_t b) {
+  writeFully(&b, 1);
+}
+
+void EthernetClient::writeFully(const char *buf) {
+  writeFully(reinterpret_cast<const uint8_t *>(buf), strlen(buf));
+}
+
+void EthernetClient::writeFully(const char *buf, size_t size) {
+  writeFully(reinterpret_cast<const uint8_t *>(buf), size);
+}
+
+void EthernetClient::writeFully(const uint8_t *buf, size_t size) {
+  // Don't use client.connected() as the "connected" check because
+  // that will return true if there's data available, and this loop
+  // does not check for data available.
+  while (size > 0 && static_cast<bool>(*this)) {
+    size_t written = write(buf, size);
+    size -= written;
+    buf += written;
+  }
+}
+
 size_t EthernetClient::write(uint8_t b) {
-  if (!(*this)) {
+  if (!static_cast<bool>(*this)) {
     return 0;
   }
   const auto &state = conn_->state;
@@ -218,24 +257,24 @@ size_t EthernetClient::write(uint8_t b) {
     return 0;
   }
 
+  size_t sndBufSize = tcp_sndbuf(state->pcb);
+  if (sndBufSize == 0) {  // Possibly flush if there's no space
+    tcp_output(state->pcb);
+    sndBufSize = tcp_sndbuf(state->pcb);
+  }
+
   size_t written = 0;
-
-  if (tcp_sndbuf(state->pcb) >= 1) {
-    if (tcp_write(state->pcb, &b, 1, TCP_WRITE_FLAG_COPY) != ERR_OK) {
-      return 0;
+  if (sndBufSize >= 1) {
+    if (tcp_write(state->pcb, &b, 1, TCP_WRITE_FLAG_COPY) == ERR_OK) {
+      written = 1;
     }
-    written = 1;
   }
-
-  // Potentially wait for some data to flush
-  if (conn_->connected && tcp_sndbuf(state->pcb) == 0) {
-    EthernetClass::loop();
-  }
+  EthernetClass::loop();  // Loop to allow incoming TCP data
   return written;
 }
 
 size_t EthernetClient::write(const uint8_t *buf, size_t size) {
-  if (!(*this)) {
+  if (!static_cast<bool>(*this)) {
     return 0;
   }
 
@@ -244,26 +283,29 @@ size_t EthernetClient::write(const uint8_t *buf, size_t size) {
     return 0;
   }
 
-  if (size > UINT16_MAX) {
-    size = UINT16_MAX;
+  if (size == 0) {
+    EthernetClass::loop();  // Loop to allow incoming TCP data
+    return 0;
   }
+
   size_t sndBufSize = tcp_sndbuf(state->pcb);
+  if (sndBufSize == 0) {  // Possibly flush if there's no space
+    tcp_output(state->pcb);
+    sndBufSize = tcp_sndbuf(state->pcb);
+  }
   size = std::min(size, sndBufSize);
   if (size > 0) {
     if (tcp_write(state->pcb, buf, size, TCP_WRITE_FLAG_COPY) != ERR_OK) {
-      return 0;
+      size = 0;
     }
   }
 
-  // Potentially wait for some data to flush
-  if (conn_->connected && tcp_sndbuf(state->pcb) == 0) {
-    EthernetClass::loop();
-  }
+  EthernetClass::loop();  // Loop to allow incoming TCP data
   return size;
 }
 
 int EthernetClient::availableForWrite() {
-  if (!(*this)) {
+  if (!static_cast<bool>(*this)) {
     return 0;
   }
   const auto &state = conn_->state;
@@ -271,27 +313,26 @@ int EthernetClient::availableForWrite() {
     return 0;
   }
 
-  // Potentially wait for some data to flush (known to be connected)
+  // Maybe flush
   if (tcp_sndbuf(state->pcb) == 0) {
-    EthernetClass::loop();
+    tcp_output(state->pcb);
   }
+
+  EthernetClass::loop();  // Loop to allow incoming TCP data
   return tcp_sndbuf(state->pcb);
 }
 
 void EthernetClient::flush() {
-  if (!(*this)) {
+  if (!static_cast<bool>(*this)) {
     return;
   }
   const auto &state = conn_->state;
   if (state == nullptr) {
     return;
   }
-  tcp_output(state->pcb);
 
-  // Potentially wait for some data to flush
-  if (conn_->connected && tcp_sndbuf(state->pcb) == 0) {
-    EthernetClass::loop();
-  }
+  tcp_output(state->pcb);
+  EthernetClass::loop();  // Loop to allow incoming TCP data
 }
 
 // --------------------------------------------------------------------------
@@ -386,15 +427,17 @@ int EthernetClient::read(uint8_t *buf, size_t size) {
     conn_ = nullptr;
     return 0;
   }
-  if (size == 0) {
-    return 0;
-  }
 
   const auto &state = conn_->state;
   if (state == nullptr) {
     return 0;
   }
+
   EthernetClass::loop();  // Allow data to come in
+  if (size == 0) {
+    return 0;
+  }
+
   // NOTE: loop() requires a re-check of the state
   if (!isAvailable(state)) {
     return 0;