+++ src/QNEthernetServer.h
@@ -12,6 +12,7 @@
 
 #include <Print.h>
 #include <Server.h>
+
 #include "QNEthernetClient.h"
 
 namespace qindesign {
@@ -22,7 +23,13 @@ class EthernetServer final : public Server {
   EthernetServer(uint16_t port);
   ~EthernetServer();
 
+  // Start listening on the server port. This calls begin(false).
   void begin() override;
+
+  // Start listening on the server port and set the SO_REUSEADDR socket option
+  // according to the `reuse` parameter.
+  void begin(bool reuse);
+
   bool end();
 
   EthernetClient accept() const;