+++ src/QNEthernetUDP.cpp
@@ -12,10 +12,13 @@
 #include <elapsedMillis.h>
 #include <lwip/dns.h>
 #include <lwip/igmp.h>
+#include <lwip/ip.h>
 #include <lwip/opt.h>
+
+#include "QNDNSClient.h"
 #include "QNEthernet.h"
 
-extern const int kMTU;
+extern const size_t kMTU;
 
 namespace qindesign {
 namespace network {
@@ -27,21 +30,6 @@ const size_t kMaxUDPSize = kMTU - 8 - 20;
 // DNS lookup timeout.
 static constexpr uint32_t kDNSLookupTimeout =
     DNS_MAX_RETRIES * DNS_TMR_INTERVAL;
-
-void EthernetUDP::dnsFoundFunc(const char *name, const ip_addr_t *ipaddr,
-                               void *callback_arg) {
-  if (callback_arg == nullptr || ipaddr == nullptr) {
-    return;
-  }
-
-  EthernetUDP *udp = reinterpret_cast<EthernetUDP *>(callback_arg);
-
-  // Also check the host name in case there was some previous request pending
-  if (udp->lookupHost_ == name) {
-    udp->lookupIP_ = ipaddr->addr;
-    udp->lookupFound_ = true;
-  }
-}
 
 void EthernetUDP::recvFunc(void *arg, struct udp_pcb *pcb, struct pbuf *p,
                            const ip_addr_t *addr, u16_t port) {
@@ -89,11 +77,20 @@ EthernetUDP::~EthernetUDP() {
 }
 
 uint8_t EthernetUDP::begin(uint16_t localPort) {
+  return begin(localPort, false);
+}
+
+uint8_t EthernetUDP::begin(uint16_t localPort, bool reuse) {
   if (pcb_ == nullptr) {
     pcb_ = udp_new();
   }
   if (pcb_ == nullptr) {
     return false;
+  }
+
+  // Try to bind
+  if (reuse) {
+    ip_set_option(pcb_, SOF_REUSEADDR);
   }
   if (udp_bind(pcb_, IP_ANY_TYPE, localPort) != ERR_OK) {
     return false;
@@ -144,12 +141,13 @@ int EthernetUDP::parsePacket() {
   packet_ = inPacket_;
   inPacket_.clear();
 
+  EthernetClass::loop();  // Allow the stack to move along
+
   if (packet_.size() > 0) {
     packetPos_ = 0;
     return packet_.size();
   } else {
     packetPos_ = -1;
-    EthernetClass::loop();  // Allow the stack to move along
     return 0;
   }
 }
@@ -228,28 +226,11 @@ int EthernetUDP::beginPacket(IPAddress ip, uint16_t port) {
 }
 
 int EthernetUDP::beginPacket(const char *host, uint16_t port) {
-  ip_addr_t addr;
-  lookupHost_ = host;
-  lookupIP_ = INADDR_NONE;
-  lookupFound_ = false;
-  switch (dns_gethostbyname(host, &addr, &dnsFoundFunc, this)) {
-    case ERR_OK:
-      return beginPacket(addr.addr, port);
-    case ERR_INPROGRESS: {
-      elapsedMillis timer;
-      while (lookupIP_ == INADDR_NONE && timer < kDNSLookupTimeout) {
-        // NOTE: Depends on Ethernet loop being called from yield()
-        yield();
-      }
-      if (lookupFound_) {
-        return beginPacket(lookupIP_, port);
-      }
-      return false;
-    }
-    case ERR_ARG:
-    default:
-      return false;
+  IPAddress ip;
+  if (!DNSClient::getHostByName(host, ip, kDNSLookupTimeout)) {
+    return false;
   }
+  return beginPacket(ip, port);
 }
 
 int EthernetUDP::endPacket() {