+++ src/lwipopts.h
@@ -38,7 +38,7 @@
 // #define MEMP_NUM_PBUF                      16
 // #define MEMP_NUM_RAW_PCB                   4
 // Add one to MEMP_NUM_UDP_PCB for mDNS:
-#define MEMP_NUM_UDP_PCB                   5
+#define MEMP_NUM_UDP_PCB                   8
 #define MEMP_NUM_TCP_PCB                   8
 // #define MEMP_NUM_TCP_PCB_LISTEN            8
 // #define MEMP_NUM_TCP_SEG                   16
@@ -53,10 +53,10 @@
 //    (LWIP_IPV6 * (1 + LWIP_IPV6_REASS + LWIP_IPV6_MLD + LWIP_IPV6_DHCP6)))*/
 // Add 3 for mDNS. This value seems to work:
 #define MEMP_NUM_SYS_TIMEOUT ((LWIP_NUM_SYS_TIMEOUT_INTERNAL) + (3))
-#define MEMP_NUM_NETBUF                    0
-#define MEMP_NUM_NETCONN                   0
+// #define MEMP_NUM_NETBUF                    2
+// #define MEMP_NUM_NETCONN                   4
 // #define MEMP_NUM_SELECT_CB                 4
-#define MEMP_NUM_TCPIP_MSG_API             0
+// #define MEMP_NUM_TCPIP_MSG_API             8
 // #define MEMP_NUM_TCPIP_MSG_INPKT           8
 // #define MEMP_NUM_NETDB                     1
 // #define MEMP_NUM_LOCALHOSTLIST             1
@@ -339,6 +339,7 @@
 // Hook options
 
 // Debugging options
+// #define LWIP_DEBUG
 // #define LWIP_DBG_MIN_LEVEL              LWIP_DBG_LEVEL_ALL
 // #define LWIP_DBG_TYPES_ON               LWIP_DBG_ON
 // #define ETHARP_DEBUG                    LWIP_DBG_OFF
@@ -412,5 +413,6 @@
 #define LWIP_MDNS_RESPONDER LWIP_UDP
 // #define MDNS_RESP_USENETIF_EXTCALLBACK LWIP_NETIF_EXT_STATUS_CALLBACK
 #define MDNS_MAX_SERVICES   3
+// #define MDNS_DEBUG                       LWIP_DBG_OFF
 
 #endif  // LWIPTEENSY_LWIPOPTS_H_